% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{forint}
\alias{forint}
\title{Formats numbers as Hungarian Forint}
\usage{
forint(x)
}
\arguments{
\item{x}{number of forint}
}
\value{
formatted string
}
\description{
Formats numbers as Hungarian Forint
}
\examples{
forint(42)
forint(100246.73)
}
