% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bitcoin_price.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Gets the price of BTC in USD}
\usage{
get_bitcoin_price()
}
\value{
number
}
\description{
Gets the price of BTC in USD
}
\examples{
get_bitcoin_price()
}
