% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_rates.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Look up the historical exchange rate between a base and target currency. Defaults to USD - HUF.}
\usage{
get_exchange_rates(
  base_currency = "USD",
  target_currency = "HUF",
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  retried = 0
)
}
\arguments{
\item{base_currency}{currency symbol}

\item{target_currency}{currency symbol}

\item{start_date}{date}

\item{end_date}{date}

\item{retried}{number of times the function has been retried}
}
\value{
\code{data.table} object
}
\description{
Look up the historical exchange rate between a base and target currency. Defaults to USD - HUF.
}
\examples{
get_exchange_rates()
get_usdhufs("USD", "EUR", start_date = "2021-04-18", end_date = "2021-05-18")
}
