% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_usdhuf}
\alias{get_usdhuf}
\title{Gets value of 1 USD in Hungarian Forints}
\usage{
get_usdhuf(retried = 0)
}
\arguments{
\item{retried}{number of times the function has been retried}
}
\value{
number
}
\description{
Gets value of 1 USD in Hungarian Forints
}
\examples{
get_usdhuf()
}
